[35] Block import logic (#63)
* Block import logic

* Spotless

* Replaced finalization state with finalization state provider

* Added tests for new JsonRpcExecutionLayerManager methods

* Some refactoring EngineApiBlockCreator, ExecutionLayerManager, added functionality to QbftBlockImporter, change of ProposerSelector

* Broke down block importer. Improved test coverage

* Spotless

* Minor improvement

* Minor rename

* Reverted improper rename

* Reverting the rename

* Argument rename

* Moved in memory beacon chain to be reusable

* Minor improvements

+++ maru/consensus/src/main/kotlin/maru/consensus/qbft/QbftBlockCreator.kt
@@ -68,7 +68,7 @@ class QbftBlockCreator(
       BeaconBlockBody(latestBeaconBlock.commitSeals, executionPayload)
     val proposer =
       proposerSelector.selectProposerForRound(
-        ConsensusRoundIdentifier((parentBeaconBlockHeader.number + 1UL).toLong(), round.toInt()),
+        ConsensusRoundIdentifier((parentBeaconBlockHeader.number + 1UL).toLong(), round),
       )
     val stateRootBlockHeader =
       BeaconBlockHeader(
